feat: add `FormsPagesContainer` component
Add `FormsPagesContainer` component which is the container that carries
most parts of "log in" or "sign up" pages and just needs to take the form
of the page as child to build a complate "log in" or "sign up" page.

Add `ocen-bg.jpg` and `tree-bg.jpg` images to `images` folder to use them as
background images in `FormsPagesContainer` component.

Declare a module type for "jpg" images in `Globals.d.ts` file.

Make `components/OptionsMenu/index.ts` file export an object carries
the menu component itself and option component that acts as menu item
of menu component.

Cancel setting `MenuOption` component as member inside `OptionsMenu` component.